saro-persona-rbac/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic[email]==2.9.0
PyJWT==2.9.0
python-multipart==0.0.12
httpx==0.27.0
pytest==8.3.0
